Add outline to pdf documents
Ideally this would be handled by Chrome during printing. However,
https://bugs.chromium.org/p/chromium/issues/detail?id=840455 is not
implemented yet (and cannot rely on metadata extracted from the asciidoc
format directly).

Therefore this implements it by introducing some kind of post-processing
using `pdf-lib`. If the `:toc:` property is set, the `.adoc` file is
scanned for sections (respecting the `:toclevels:` attribute) and an
outline is generated.

This only works if a ToC is also generated within the document (or
better: links exist to each section), because otherwise Chrome would not
generate the necessary `Dests` fields within the PDF.

Unfortunately, Chrome also has some bugs regarding Umlaute in anchors,
leading to omission of the relevant `Dests` fields. Therefore a warning
is printed if a anchor cannot be located in the `Dests` field of the PDF
catalog.

Based upon
https://gitlab.pagedmedia.org/tools/pagedjs-cli/commit/df0d10bd1bb12d1e2077323e1fb5eec75da35a1e
which itself is based on @Dopding's comment at:
https://github.com/Hopding/pdf-lib/issues/127#issuecomment-502450179